+++ src/controllers/v1/user/delete_user.ts
@@ -20,10 +20,12 @@ const deleteUser = async (req: Request, res: Response): Promise<void> => {
       .lean()
       .exec();
     const publicIds = blogs.map(({ banner }) => banner.publicId);
-    await cloudinary.api.delete_resources(publicIds);
-    logger.info('Multiple blog banners deleted from cloudinary', {
-      publicIds,
-    });
+    if(publicIds.length > 0) {
+      await cloudinary.api.delete_resources(publicIds);
+      logger.info('Multiple blog banners deleted from cloudinary', {
+        publicIds,
+      });
+    }
 
     await Blog.deleteMany({ author: userId });
     logger.info('Multiple blogs deleted', {
